refactor about section
