+++ index.ts
@@ -1,7 +1,6 @@
 import express, { Express } from 'express';
 import indexRoutes from './routes/index.js';
 import catRoutes from './routes/cats.js';
-import ageRoutes from './routes/age.js';
 
 const app: Express = express();
 const PORT: number = parseInt(process.env.PORT || '3000', 10);
@@ -14,7 +13,6 @@ app.use(express.json());
 // Routes
 app.use('/', indexRoutes);
 app.use('/api/cats', catRoutes);
-app.use('/api/age', ageRoutes);
 
 app.listen(PORT, () => {
     console.log(`Server running at http://localhost:${PORT}`);
