+++ routes/cats.ts
@@ -10,7 +10,30 @@ interface Cat {
     fact: string;
 }
 
-const cats: Cat[] = [
+interface Age {
+    years: number;
+    months: number;
+}
+
+function calculateAge(birthDate: Date): Age {
+    const today = new Date();
+
+    let years = today.getFullYear() - birthDate.getFullYear();
+    let months = today.getMonth() - birthDate.getMonth();
+
+    if (months < 0 || (months === 0 && today.getDate() < birthDate.getDate())) {
+        years--;
+        months += 12;
+    }
+
+    if (today.getDate() < birthDate.getDate()) {
+        months--;
+    }
+
+    return { years, months };
+}
+
+const cats: Array<Cat> = [
     {
         id: 1,
         name: 'Groucho',
@@ -26,6 +49,31 @@ const cats: Cat[] = [
         fact: 'Purrs louder than a motor!'
     }
 ];
+
+router.get('/about', (_req: Request, res: Response): void => {
+    const birthDate = new Date('2021-07-02');
+    const age = calculateAge(birthDate);
+
+    let ageText = `${age.years} year${age.years !== 1 ? 's' : ''}`;
+    if (age.months > 0) {
+        ageText += ` and ${age.months} month${age.months !== 1 ? 's' : ''}`;
+    }
+    ageText += ' old';
+
+    const aboutItems: Array<string> = [
+        '<strong>Gotcha day:</strong> November 10, 2021',
+        'Adopted from <a class="text-orange-600 hover:underline" href="https://www.sfspca.org/" target="_blank">SF SPCA</a>',
+        'The names, Groucho and Chica, are from the shelter',
+        'They were in the same shelter room and bonded together',
+        'Not siblings, but they have similar fur patterns',
+        'Both are American Shorthair cats',
+        `Both are born on July 2, 2021 (currently ${ageText})`,
+        'They love to play and sleep together'
+    ];
+
+    const html = aboutItems.map((item) => `<li>${item}</li>`).join('');
+    res.send(html);
+});
 
 router.get('/:id', (req: Request, res: Response): void => {
     const catId = parseInt(req.params.id);
